feat: completed rewiring all cats and all requests
