+++ components/CatNotifier.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CatNotifer: React.FC = () => {
+  return <div className="copy-notifer">Copied Cat Id</div>;
+};
+
+export default React.memo(CatNotifer);
